Rename protocols to protocol mock.
