+++ credit/build.gradle
@@ -20,7 +20,7 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-webflux'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
-	implementation project(':protocols')
+	implementation project(':protocolmock')
 }
 
 test {
