+++ protocolmock/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id "io.freefair.lombok" version "5.3.0"
 }
 
-group 'org.example'
+group 'de.upb.crypto'
 version '1.0-SNAPSHOT'
 
 repositories {
