+++ protocolmock/src/main/java/de/upb/crypto/incentive/protocolmock/deduct/DeductRequest.java
@@ -1,6 +1,6 @@
-package de.upb.crypto.incentive.protocols.deduct;
+package de.upb.crypto.incentive.protocolmock.deduct;
 
-import de.upb.crypto.incentive.protocols.model.Token;
+import de.upb.crypto.incentive.protocolmock.model.Token;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
