+++ protocolmock/src/main/java/de/upb/crypto/incentive/protocolmock/issue/IssueResponse.java
@@ -1,6 +1,6 @@
-package de.upb.crypto.incentive.protocols.credit;
+package de.upb.crypto.incentive.protocolmock.issue;
 
-import de.upb.crypto.incentive.protocols.model.Token;
+import de.upb.crypto.incentive.protocolmock.model.Token;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CreditResponse {
+public class IssueResponse {
     private int id;
     private Token token;
 }
