+++ protocolmock/src/main/java/de/upb/crypto/incentive/protocolmock/model/Reward.java
@@ -1,12 +1,13 @@
-package de.upb.crypto.incentive.protocols.model;
+package de.upb.crypto.incentive.protocolmock.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
-public class Token {
+@AllArgsConstructor
+public class Reward {
+    private String name;
     private int value;
 }
